test(zwave_js): consolidate shared provider fixtures into conftest
